# acp-writer/api/openapi.yaml
openapi: 3.1.0
info:
  title: ACP Writer API
  description: |
    Composes patient-specific, FHIR-compliant care plans by combining clinical
    decision logic (DMN) with patient data. Manages an internal decision engine
    (Drools/Kogito) and knowledge base (vector store).

    Callers provide patient data directly — the acp-writer does not query FHIR
    servers. The decision engine and vector store are internal implementation
    details hidden behind this API.

    Contract version: 1.0
  version: 0.2.0
  license:
    name: Apache 2.0
    url: https://www.apache.org/licenses/LICENSE-2.0
servers:
  - url: http://localhost:8082
    description: Local development
paths:
...
  /api/v1/careplans:
    post:
      operationId: generateCarePlan
      summary: Generate a care plan from patient data
      description: |
        Accepts patient data as a FHIR Bundle or IPS document. Internally
        extracts relevant clinical data, invokes applicable DMN decision
        models, retrieves relevant recommendations from the knowledge base,
        and assembles a FHIR CarePlan Bundle.

        The caller is responsible for providing patient data — this service
        does not query FHIR servers directly.
      tags: [careplans]
      requestBody:
        required: true
        content:
          application/fhir+json:
            schema:
              type: object
              description: FHIR Bundle or IPS document containing patient data
      responses:
        "201":
          description: Care plan generated
          content:
            application/fhir+json:
              schema:
                type: object
                description: FHIR Bundle containing CarePlan, Goal, and activity resources
        "400":
          description: Invalid patient data
        "422":
          description: Unable to generate care plan (e.g., no applicable decision models)

    get:
      operationId: listCarePlans
      summary: List generated care plans
      tags: [careplans]
      parameters:
        - name: patient
          in: query
          schema:
            type: string
          description: Filter by patient reference
        - name: status
          in: query
          schema:
            type: string
            enum: [draft, approved, rejected]
          description: Filter by care plan status
      responses:
        "200":
          description: List of care plans
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/CarePlanSummary"
  /api/v1/careplans/{careplanId}:
    parameters:
      - name: careplanId
        in: path
        required: true
        schema:
          type: string

    get:
      operationId: getCarePlan
      summary: Retrieve a generated care plan
      tags: [careplans]
      responses:
        "200":
          description: Care plan bundle
          content:
            application/fhir+json:
              schema:
                type: object
                description: FHIR Bundle containing CarePlan and related resources
        "404":
          description: Care plan not found
  /api/v1/careplans/{careplanId}/status:
    put:
      operationId: updateCarePlanStatus
      summary: Update care plan status
      tags: [careplans]
      parameters:
        - name: careplanId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CarePlanStatusUpdate"
      responses:
        "200":
          description: Status updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CarePlanSummary"
        "404":
          description: Care plan not found

  # --- Health ---
...
components:
  schemas:
...
    CarePlanSummary:
      type: object
      properties:
        id:
          type: string
        patient_reference:
          type: string
        title:
          type: string
        status:
          type: string
          enum: [draft, approved, rejected]
        generated_at:
          type: string
          format: date-time
        decision_models_used:
          type: array
          items:
            type: string
    CarePlanStatusUpdate:
      type: object
      required: [status]
      properties:
        status:
          type: string
          enum: [approved, rejected]
        reason:
          type: string
        clinician:
          type: string
          description: Legacy display-name shortcut for the verifier; prefer `reviewer`.
        reviewer:
          $ref: "#/components/schemas/ReviewerRef"
    ReviewerRef:
      type: object
      description: >
        Identity of the human verifier recorded on the AI-Provenance when the
        plan is approved. Optional request override; falls back to the
        ACP_REVIEWER_* config defaults. A SMART-on-FHIR launch would populate
        this from the token's fhirUser claim. Mirrors the ReviewerRef schema in
        bff-openapi.yaml — the BFF and REST specs are independent documents, so
        each defines it once (no cross-file $ref).
      required: [display]
      properties:
        display:
          type: string
          description: Human-readable verifier name.
        reference:
          type: string
          description: 'FHIR reference to the verifier, e.g. "Practitioner/demo-clinician".'
        identifierSystem:
          type: string
          description: Identifier system for the verifier (paired with identifierValue).
        identifierValue:
          type: string
          description: Identifier value for the verifier (paired with identifierSystem).

    # --- Health ---
...